bigquery-importer/requirements.txt
pandas==2.1.4
faker==21.0.0
google-cloud-storage==2.10.0
google-cloud-bigquery==3.13.0
python-dotenv==1.0.0
pyarrow==14.0.2
db-dtypes==1.2.0
